resolved cyclic dependency
